pyproject.toml
[project]
name = "scriber"
version = "0.1.0"
description = "Offline transcription for qualitative researchers"
requires-python = ">=3.12"
dependencies = [
    "pyqt6>=6.7.0",
    "faster-whisper>=1.0.0",
    "pyannote.audio>=4.0.0",
    "av>=12.0.0",
    "platformdirs>=4.0.0",
    "numpy>=1.26.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

[project.optional-dependencies]
mlx = [
    "mlx>=0.16.0",
    "mlx-whisper>=0.3.0",
]
dev = [
    "pyinstaller>=6.0.0",
    "pytest>=8.0.0",
]

[project.scripts]
scriber = "scriber.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pyinstaller>=6.0.0",
    "pytest>=8.0.0",
]
